% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.header.R
\name{add.header}
\alias{add.header}
\title{Addin snippet function to add header comment to a current opened file}
\usage{
add.header()
}
\value{
Inserts header content for file
}
\description{
Shorthand to add header comment
}
\examples{
if(interactive())
add.header()
}
