% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.header.R
\name{add.sect.comment}
\alias{add.sect.comment}
\title{Addin snippet function to custom section comment}
\usage{
add.sect.comment()
}
\value{
Inserts section comment content for file
}
\description{
Shorthand to add section comment to current file
}
\examples{
if(interactive())
add.sect.comment()
}
