% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add.snippet.clear}
\alias{add.snippet.clear}
\title{Snippet R function to clear console and set directory}
\usage{
add.snippet.clear()
}
\value{
Inserts code to clear console
}
\description{
Shorthand to add clear console code to current file
}
\examples{
if(interactive())
add.snippet.clear()
}
