% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate.R
\name{ai.duplicate}
\alias{ai.duplicate}
\title{Prompt guided duplication and editing of files}
\usage{
ai.duplicate(file = NULL, new.name = NULL, open = TRUE)
}
\arguments{
\item{file}{file to duplicate}

\item{new.name}{OPTIONAL.name of new file}

\item{open}{open file after duplication}
}
\value{
duplicated files with edited texts
}
\description{
AI like duplication and editing of files
}
\examples{
\donttest{
if(interactive()){
file1s <- paste0(tempfile(),".R")
writeLines("message(
'Sample items: farm, shinyappstore, rpkg'
)", file1s)
ai.duplicate(file1s,'file2.R')
}
}


}
