% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedR.R
\name{archivedPkg}
\alias{archivedPkg}
\title{Listing of all CRAN archived R packages}
\usage{
archivedPkg(
  startsWith = letters,
  inc.date = TRUE,
  as = c("data.frame", "vector")
)
}
\arguments{
\item{startsWith}{one letter that the package name starts with eg. a, e, f}

\item{inc.date}{should archive date be included in the result}

\item{as}{return result as data frame or as vector}
}
\value{
a data frame container listing of all archived R packages
}
\description{
Retrieve a list of all currently archived R packages and their archive date
}
\note{
The startsWith argument should be one letter and should be in lowercase
}
\examples{
# Task 1: get archived R packages with names beginning with A
head(archivedPkg(startsWith = "a"))


# Task 2: return the packages from Task 1 without including latest archive date
res.dt2 <- archivedPkg(startsWith = "a", inc.date = FALSE)
res.dt2[1:10,]

# Task 3: return the results from Task 2 as a vector
res.dt3 <- archivedPkg(startsWith = "a", inc.date = FALSE, as = "vector")
res.dt2[1:10]

# Task 4: return the archived packages beginning with Y
# Note that startsWith should be lowercase
head(archivedPkg(startsWith = "y"))


}
