% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean}
\alias{clean}
\title{Clear environment, clear console, set work directory and load files}
\usage{
clean(setwd = NULL, source = c(), load = c())
}
\arguments{
\item{setwd}{OPTIONAL. set working directory}

\item{source}{OPTIONAL. source in file(s)}

\item{load}{OPTIONAL. load in Rdata file(s)}
}
\value{
cleared environment and set directory
}
\description{
Shorthand to quickly clear console, clear environment, set working directory, load files
}
