% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersec.R
\name{get_active_conn}
\alias{get_active_conn}
\title{Cyber security functions}
\usage{
get_active_conn()
}
\value{
a data frame containing the list of all connextions
}
\description{
Get all active connections
}
\examples{

# Get all connections on my system
x = get_active_conn()


}
