% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pop.R
\name{data_pop}
\alias{data_pop}
\title{Remove last n rows or column or specified elements from a data frame like array_pop in PHP}
\usage{
data_pop(., n = 1, which = c("rows", "cols"), ret = FALSE)
}
\arguments{
\item{.}{parent vector}

\item{n}{number of elements to remove}

\item{ret}{TRUE or FALSE. whether to return value instead of setting it to the parent vector}

\item{el}{vector to remove}
}
\value{
vector with elements removed
}
\description{
Shorthand to remove elements from a data frame and save as the same name
}
\examples{
data.01 <- mtcars

#task: remove 1 element from the end of the vector and set it to the vector name
data.01 #data.01 data before pop
data_pop(data.01) #does not return anything
data.01 #data.01 data updated after pop

#task: remove 5 elements from the end, but do not set it to the vector name
data.01 #data.01 vector before pop
data_pop(data.01,5, ret = TRUE) #return modified vector
data.01 #data.01 vector remains the same after pop

}
