% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_pop_filter}
\alias{data_pop_filter}
\title{Remove elements from a data based on filter}
\usage{
data_pop_filter(., remove)
}
\arguments{
\item{.}{data object}

\item{remove}{expression for filter}
}
\value{
data filtered out based on the expression
}
\description{
Shorthand to remove elements from a data frame based on filter and save as the same name
}
\examples{

data.01 <- mtcars

#task: remove all mpg > 20
data.01 #data.01 data before pop
data_pop_filter(data.01,mpg > 30 ) #computes and resaves to variable
data.01 #modified data after pop based on filter

#task: remove all multiple. remove all elements where carb  == 4 or hp > 200
data.01 #data.01 data before pop
data_pop_filter(data.01,carb  == 4 | hp > 200 ) #computes and resaves to variable
data.01 #modified data after pop based on filter

}
