% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_push.R
\name{data_push}
\alias{data_push}
\title{Add data to another data like array_push in PHP}
\usage{
data_push(., add, which = c("rows", "cols"))
}
\arguments{
\item{.}{first data set}

\item{add}{data set to add}

\item{which}{where to append the new data e.g. rows or cols}
}
\value{
the combined dataset store to a variable with the name of the first
}
\description{
Shorthand to add data to a dataset and save as the same name
}
\examples{
p1 <- data.frame(PK=1:10,ID2=1:10)
p2 <- data.frame(PK=11:20,ID2=21:30)
data_push(p1,p2,"rows")
}
