% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_shuffle}
\alias{data_shuffle}
\title{Shuffle a dataframe just like shuffle but for data frame}
\usage{
data_shuffle(., which = c("rows", "cols"))
}
\description{
Shuffle a dataframe just like shuffle but for data frame
}
