% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.8.R
\name{date3to1}
\alias{date3to1}
\title{Combine vectors to form date}
\usage{
date3to1(data, default = "Start", col.YMD = 1:3)
}
\description{
Adapted from Ecfun R package
}
\examples{

df1 <- data.frame(Year=c(NA, -1, 1971:1979),
Month=c(1:2, -1, NA, 13, 2, 12, 6:9),
Day=c(0, 0:6, NA, -1, 32) )
head(df)
DateVecS <- date3to1(df1)
}
