% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_desc_asc.R
\name{is.increasing}
\alias{is.increasing}
\alias{is.decreasing}
\title{Check is numbers in a vector is decreasing or increasing}
\usage{
is.increasing(.)

is.decreasing(.)
}
\description{
Check is numbers in a vector is decreasing or increasing
}
