% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{duplicate}
\alias{duplicate}
\title{Duplicate a file and global replace}
\usage{
duplicate(file, new.name, pattern, replacement, open = TRUE)
}
\arguments{
\item{file}{data frame to re-sample}

\item{new.name}{column to uniquely re-sample}

\item{pattern}{number of rows to return}

\item{replacement}{unique numeric value for reproducibility}

\item{open}{description}
}
\value{
data frame containing re-sampled rows from an original data frame
}
\description{
Shorthand to return a re-sample number of rows in a data frame by unique column
}
\examples{
\donttest{
duplicate('./file.R','file2.R','text1','replacement1')
}
}
