% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR.R
\name{error.out}
\alias{error.out}
\alias{\%eo\%}
\title{Error coalescing operator}
\usage{
error.out(test, alternative = "")

test \%eo\% alternative
}
\arguments{
\item{test}{an object to return}

\item{alternative}{alternative object to return}
}
\value{
value of test if error, else return value of alternative
}
\description{
Alternative return for error statements.Return alternative if the value of expression is erroneous
}
\examples{
# The following statement would produce
# error because 'stat1' does not exist

# stat1 + 1

# To prevent the statement from
# stopping the process, when can have alternative out
alt = 0
error.out(stats1 + 1, alt)

}
