% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{extract_comments}
\alias{extract_comments}
\title{Extract all comments from a file}
\usage{
extract_comments(file)
}
\arguments{
\item{file}{file to parse}
}
\value{
vector of all comments within a file
}
\description{
Vectorize all comments from the file
}
\examples{
\dontrun{
ex_file1 <- "path/file1.R"
# get all comments
cmmts <- extract_comments(ex_file1)
cmmts
}

}
