% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_by.R
\name{find_packages}
\alias{find_packages}
\title{Fetch R package based on keyword}
\usage{
find_packages(keyword)
}
\arguments{
\item{keyword}{character, the keyword to search package descriptions for.}
}
\value{
character vector of matching package names or NULL if no matches.
}
\description{
This function gets R packages based on a keyword in their description.
}
\note{
the function in its current form only searches available.packages()
}
\examples{
\donttest{
# find the list of R packages for data or machine learning

matched_pkgs <- find_packages("data")
matched_pkgs

matched_pkgs <- find_packages("machine learning")
matched_pkgs
}
}
