% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aided_chatgpt.R
\name{from_tensor_slices}
\alias{from_tensor_slices}
\title{Create Tensor-Like Slices from a Data Frame or Matrix}
\usage{
from_tensor_slices(data)
}
\arguments{
\item{data}{A data frame or matrix to be converted into slices.}
}
\value{
A list of slices, each corresponding to a row of the input data.
}
\description{
This function converts a data frame or matrix into row-wise slices, similar to TensorFlow's `from_tensor_slices()`.
}
\examples{
df <- data.frame(A = 1:3, B = c("x", "y", "z"))
from_tensor_slices(df)

}
