% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{g.mean}
\alias{g.mean}
\title{Calculate geometric mean and round}
\usage{
g.mean(num, na.rm = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{na.rm}{remove NAs from the vector}

\item{round}{round result to decimal place}
}
\value{
the geometric mean of a set of numbers
}
\description{
Calculate the geometric mean
}
\examples{
num1 <- sample(300:3000,10)
g.mean(num1)

}
