% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomImage.R
\name{genImage}
\alias{genImage}
\title{Download random images from the web}
\usage{
genImage(fp, cat, n = 1, w.px = 500, h.px = 500, ext = "jpg")
}
\arguments{
\item{fp}{Storage directory}

\item{cat}{category of image to download}

\item{n}{number of images to download}

\item{w.px}{width in pixels}

\item{h.px}{height in pixels}

\item{ext}{file extension eg jpg, png}

\item{size}{size of images to download}
}
\value{
downloaded image from a select image category
}
\description{
Retrieve n number of high-definition images by category from the web
}
\section{SOURCE}{

The random images are downloaded from www.unsplash.com
}

\section{USE CASE}{

This functionality is great for R developers trying to obtain one or more images for using in analysis or simply for building robust
web applications.
}

\examples{
# download 10 image from the nature category
genImage(fp="D:/GithubRepo/samples/randomimages",cat = "nature", n = 100)



}
