% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.mean.R
\name{geo.mean}
\alias{geo.mean}
\title{Calculate geometric mean and round}
\usage{
geo.mean(num, na.rm = TRUE, neg.rm = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{na.rm}{remove NAs from the vector}

\item{neg.rm}{remove negative values from the vector}

\item{round}{round result to decimal place}
}
\value{
the geometric mean of a set of numbers
}
\description{
Calculate the geometric mean
}
\examples{
num1 <- sample(300:3000,10)

#get the geometric mean, excluding all negatives and round to 2
geo.mean(num1)

#or
geo.mean(num1)


#get geometric mean, but round the final value to 5 decimal places
geo.mean(num1, round = 5)

}
