% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.gm.R
\name{geo.sd}
\alias{geo.sd}
\title{Calculate geometric standard deviation and round}
\usage{
geo.sd(num, na.rm = TRUE, neg.rm = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{na.rm}{remove NAs from the vector}

\item{neg.rm}{remove negative values from the vector}

\item{round}{round result to decimal place}
}
\value{
the geometric standard deviation of a set of numbers
}
\description{
Calculate the geometric standard deviation
}
\examples{
num1 <- sample(330:400,10)
sd.gm(num1,na.rm=FALSE)

}
