% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdist.R
\name{getDistribution}
\alias{getDistribution}
\title{Get distribution of a data}
\usage{
getDistribution(data, alpha = 0.05)
}
\arguments{
\item{data}{the data to check against distributions}

\item{alpha}{p-value alpha level}
}
\description{
Check a data against multiple distributions
}
