% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.8.R
\name{getGitHubRepoStart}
\alias{getGitHubRepoStart}
\title{Fetch GitHub repository Creation Date}
\usage{
getGitHubRepoStart(repo_name)
}
\arguments{
\item{repo_name}{name of the repository}
}
\value{
date of creation of repository as a character
}
\description{
Get the information of when the repository was created
}
\examples{
getGitHubRepoStart("oobianom/quickcode")

getGitHubRepoStart("cran/dplyr")

}
