% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2week.R
\name{getWeekSeq}
\alias{getWeekSeq}
\title{Derive Numeric Week Based on Dates}
\usage{
getWeekSeq(start_date, end_date)
}
\arguments{
\item{start_date}{the start date}

\item{end_date}{the end date}
}
\description{
Convert the range of date to number of weeks
}
