% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.8.R
\name{getGitRepoStart}
\alias{getGitRepoStart}
\alias{getGitRepoChange}
\title{Fetch GitHub repository Creation Date}
\usage{
getGitRepoStart(repo_name, out.format = "\%Y-\%m-\%d")

getGitRepoChange(repo_name, out.format = "\%Y-\%m-\%d")
}
\arguments{
\item{repo_name}{name of the repository}

\item{out.format}{date output format}
}
\value{
date of creation of repository as a character

date of creation of repository as a character
}
\description{
Get the information of when the repository was created

Get the information of when the repository was last changed
}
\examples{
getGitRepoStart("oobianom/quickcode")

getGitRepoStart("cran/dplyr")

getGitRepoChange(repo_name="oobianom/quickcode")

getGitRepoChange("cran/dplyr")

}
