% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervariable.R
\name{\%-\%}
\alias{\%-\%}
\title{Shorthand to subtract elements from vector or minus two lists}
\usage{
item1 \%-\% item2
}
\arguments{
\item{item1}{first vector or list}

\item{item2}{second vector or list}
}
\description{
Shorthand to subtract elements from vector or minus two lists
}
