% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in vector or array}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector entry}

\item{table}{table of items to check}
}
\value{
a boolean value to indicate if entry is present
}
\description{
Check if entry is in vector
}
\examples{
5 \%!in\% c(1:10) #FALSE
5 \%!in\% c(11:20) #TRUE

}
