% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.R
\name{\%~\%}
\alias{\%~\%}
\title{Flatten and combine strings}
\usage{
j \%~\% h
}
\arguments{
\item{j}{string or vector}

\item{h}{second string or vector}
}
\value{
combined strings
}
\description{
Flatten vectors and combine together
}
\note{
this is simply a way to further simplify paste and paste0 into a function
}
\examples{
"obi" \%~\% "is"
"you" \%~\% "are" ~ "my" \%~\% "friend"
c("none","is","better") \%~\% "for" \%~\% "now"


}
