% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{header.rmd}
\alias{header.rmd}
\title{Snippet function to add header to a current Rmd opened file}
\usage{
header.rmd()
}
\value{
Inserts header content for Rmd file
}
\description{
Shorthand to add Rmd header
}
\examples{
if(interactive())
header.rmd()
}
