% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{in.range}
\alias{in.range}
\title{Verify whether a number falls within a specified range of values.}
\usage{
in.range(value, range.min, range.max, range.vec = NULL)
}
\arguments{
\item{value}{the vector of numbers to check}

\item{range.min}{the minimum value of the range}

\item{range.max}{the maximum value of the range}

\item{range.vec}{a vector of numbers to use for the range}
}
\value{
boolean to indicate if the value or set of values are within the range
}
\description{
With a defined range of values, the function systematically examines
each provided number to determine if it falls within the specified range.
}
\details{
The described function serves the purpose of checking whether a given number
or set of numbers falls within a specified range. It operates by taking a range
of values as input and then systematically evaluates each provided number to
determine if it lies within the defined range. This function proves particularly
useful for scenarios where there is a need to assess numeric values
against predefined boundaries, ensuring they meet specific criteria or constraints.
}
\note{
The argument range.vec is utilized when users opt not to employ the range.min or
range.max arguments. If range.vec is specified, range.min and range.max are
disregarded. It's important to note that the use of range.vec is optional.
}
\examples{
# Task 1: Check if a number is within specified range
in.range(5, range.min = 3, range.max = 10) # TRUE
in.range(25, range.min = 12, range.max = 20) # FALSE

# Task 2: Check if a set of values are within a specified range
in.range(1:5, range.min = 2, range.max = 7) #
in.range(50:60, range.min = 16, range.max = 27) #


# Task 3: Check if a number is within the range of a set of numbers
in.range(5, range.vec = 1:10) # TRUE
in.range(345, range.vec = c(1001,1002,1003,1004,1005,
1006,1007,1008,1009,1010,1011,1012,1013,1014)) # FALSE

# Task 4: Check if a set of values are within the range of a set of numbers
in.range(1:5, range.vec = 4:19) #
in.range(50:60, range.vec = c(55,33,22,56,75,213,120)) #


}
