% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inc}
\alias{inc}
\title{Increment vector by value}
\usage{
inc(., add = 1)
}
\arguments{
\item{.}{vector of number(s)}

\item{add}{number to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Increment the content of a vector and resave as the vector
}
\examples{
num1 <- sample(330:400,10)
inc(num1)
inc(num1, add= 5)
}
