% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insertInText}
\alias{insertInText}
\title{Shiny app function to insert string to current file in RStudio}
\usage{
insertInText(string)
}
\arguments{
\item{string}{what to insert}
}
\value{
Inserts into current position on opened file
}
\description{
Shorthand to insert content to opened file
}
\examples{
if(interactive()){
insertInText('hello rpkg.net')
insertInText('hello world')
}
}
