% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.image.R
\name{is.image}
\alias{is.image}
\title{Is file name extension(s) an image}
\usage{
is.image(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is an image
}
\description{
Check if one or multiple file name entry is an image
}
\examples{
img.1 = "fjk.jpg"
is.image(img.1)

img.0 = "fjk.bbVG"
is.image(img.0)

img.2 = "fjk.bmp"
is.image(img.2)

img.3 = "fjk.SVG"
is.image(img.3)

# a vector of file names
v = c("logo.png", "business process.pdf",
"front_cover.jpg", "intro.docx",
"financial_future.doc", "2022 buybacks.xlsx")
(output = is.image(v))

}
