% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{lastwd}
\alias{lastwd}
\title{Go back to previous directory}
\usage{
lastwd()
}
\value{
the previous directory
}
\description{
Navigate to previous working directory if the setwd was called by clean() or refresh() function
}
\note{
In order to use this function to retrieve the last/previous working directory, you must use the quickcode::clean() or quickcode::refresh() function to set the working directory.
}
\examples{
lastwd()
}
