% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math.qt}
\alias{math.qt}
\title{Corresponding quantile for confident intervals}
\usage{
math.qt(ci = 0.9)
}
\arguments{
\item{ci}{confident interval eg. 0.9 for 90% confident intervals}
}
\value{
a vector of two numeric values for the quantile based on the confident interval chosen
}
\description{
Compute the corresponding quantile given confident interval bounds
}
\examples{

# Get the bounds for 90\% confident intervals
math.qt(0.9)

# Get the bounds for 95\% confident intervals
# use the bounds to obtain quartile
values = number(100)
ci = math.qt(0.95)
getquart = quantile(values, probs = ci)
print(getquart)



}
