% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean.gm}
\alias{mean.gm}
\title{Calculate geometric mean and round}
\usage{
\method{mean}{gm}(num, rm.na = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{rm.na}{remove NAs from the vector}

\item{round}{round result to decimal place}
}
\value{
the geometric mean of a set of numbers
}
\description{
Calculate the geometric mean
}
\examples{
num1 <- sample(300:3000,10)
mean.gm(num1)

}
