% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{minus}
\alias{minus}
\title{Decrease vector by value}
\usage{
minus(., n = 1)
}
\arguments{
\item{.}{vector of number(s)}

\item{add}{number to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
decrement the content of a vector and resave as the vector
}
\examples{
num1 <- sample(330:400,10)
num1

# increment num1 by 1
num1 #before increment
minus(num1)
num1 #after increment

# decrease num1 by 5
minus #before increment
minus(num1, add= 5)
num1 #after increment
}
