% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mix.colors}
\alias{mix.colors}
\title{Mix or Blend two or more colors}
\usage{
mix.colors(color, type = 1, alpha = 1)
}
\arguments{
\item{color}{color vector e.g see example}

\item{type}{return type of the output}

\item{alpha}{alpha or opacity of the resulting color}
}
\value{
hex for the combined color
}
\description{
Combine colors to generate a new color
}
\examples{
# color vector
colvec <- c("red","blue","violet","green")

# just one color
mix.colors(colvec[1],type = 1, alpha = 1)

# add two colors
mix.colors(colvec[1:2],type = 1, alpha = 1)

# add three colors
mix.colors(colvec[1:3],type = 1, alpha = 1)


# return type = 2

# just one color
mix.colors(colvec[1],type = 2, alpha = 1)

# add two colors
mix.colors(colvec[1:2],type = 2, alpha = 1)

# add three colors
mix.colors(colvec[1:3],type = 2, alpha = 1)


# opacity or alpha  0.5

# just one color
mix.colors(colvec[1],type = 1, alpha = 0.5)

# add two colors
mix.colors(colvec[1:2],type = 1, alpha = 0.5)

# add three colors
mix.colors(colvec[1:3],type = 1, alpha = 0.5)

}
