% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_compHist.R
\name{mix.cols.btw}
\alias{mix.cols.btw}
\title{Mix or Blend colors between two colors}
\usage{
mix.cols.btw(colors, max = 20, alpha = 1, preview = F)
}
\arguments{
\item{colors}{the vector of two colors}

\item{max}{maximum number of colors to blend between}

\item{alpha}{alpha for the new color blends}

\item{preview}{LOGICAL. preview all color generated}
}
\description{
Mix or blend multiple colors between two colors
}
\examples{
# simple mix
mix.cols.btw(c("red","brown"))

# also preview after mixing
mix.cols.btw(c("red","green"),alpha = 0.2, preview = T)

mix.cols.btw(c("red","purple","yellow","gray"),alpha = 0.2, preview = T)



}
