% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{multiply}
\alias{multiply}
\title{Multiple a vector of numeric values}
\usage{
multiply(..., na.rm = FALSE)
}
\arguments{
\item{...}{the numeric values to multiply}

\item{na.rm}{remove NAs}
}
\value{
multiple of all content
}
\description{
Multiple all the content of a vector
}
\examples{
# multiply 1 number
# returns error
# multiply(0)

# vector of numbers
numvec <- number(10, max.digits = 3)
numvec

# multiply 2 numbers
multiply(numvec[1:2])
multiply(numvec[4], numvec[5])
multiply(a = 4, b = 5)

# multiply 5 numbers
multiply(numvec[1:5])
multiply(11, 15, 12, 14, 13)
multiply(a = 4, b = 22, c = 44, d = 9, u = 10)

}
