% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_filter.R
\name{mutate_filter}
\alias{mutate_filter}
\title{Mutate a subset of dataset, and reattach it to the dataset}
\usage{
mutate_filter(., filter, ...)
}
\description{
Mutate a subset of dataset, and reattach it to the dataset
}
