% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{ndecimal}
\alias{ndecimal}
\title{Count the number of decimal places}
\usage{
ndecimal(num)
}
\arguments{
\item{num}{a numeric value}
}
\description{
Count the number of decimal places
}
\examples{
#example 1
ndecimal(9.000322)

#example 2
ndecimal(34)

#example 3
ndecimal(45.)

}
