% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superVariable.R
\name{newSuperVar}
\alias{newSuperVar}
\title{Create and Use a super variable}
\usage{
newSuperVar(..., value = 1, class = list, lock = TRUE)
}
\arguments{
\item{...}{variable name super variable}

\item{value}{value of the variable}

\item{class}{class of variable}

\item{lock}{lock variable to change}
}
\value{
no visible return, but variable is created and stored
}
\description{
Create a variable that supercedes other variables and has various functionalities
}
\examples{
# create a variable to store dataset that should not be altered
newSuperVar(mtcardf, value = mtcars)


}
