% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aided_chatgpt.R
\name{normalize.vector}
\alias{normalize.vector}
\title{Normalize a Numeric Vector to the Range [0, 1]}
\usage{
normalize.vector(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector with values scaled to the range [0, 1].
        If all values are identical, all values are set to 0.
}
\description{
This function normalizes a numeric vector so that all values are scaled to the range [0, 1].
}
\examples{
# Example with a numeric vector
normalize.vector(c(1, 2, 3, 4, 5))

# Example with a vector containing identical values
normalize.vector(c(3, 3, 3))

}
