% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.data.R
\name{not.data}
\alias{not.data}
\title{Not a data}
\usage{
not.data(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is a data table
}
\description{
Opposite of is.data.frame(). Check if entry is not a data object
}
\examples{
test.dt <- data.frame(ID=1:200,Type="RPKG.net")
test.notenv <- list(t=1)

is.data.frame(test.dt) # TRUE
not.data(test.dt) # FALSE

not.data(test.notenv) # TRUE
if(not.data(test.dt)) message("yes") # NULL

}
