% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not.empty}
\alias{not.empty}
\title{Not empty}
\usage{
not.empty(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is empty
}
\description{
Check if entry is not empty
}
\examples{
not.empty("empty") # TRUE
not.empty('') # FALSE
not.empty(NULL) # logical(0)
if(not.empty('')) message("yes") # NULL
}
