% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not.environment}
\alias{not.environment}
\title{Not an environment}
\usage{
not.environment(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is an environment
}
\description{
Check if entry is not an environment object
}
\examples{
test.env <- new.env()
test.notenv <- list(t=1)
not.environment(test.env) # FALSE
not.environment(test.notenv) # TRUE
if(not.environment(test.notenv)) message("yes") # yes

}
