% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.empty.R
\name{not.exists}
\alias{not.exists}
\title{Not exists}
\usage{
not.exists(x)
}
\arguments{
\item{x}{object}
}
\value{
a boolean value to indicate if entry does not exists
}
\description{
Check if object does not exists
}
\examples{
go = 7
not.exists("exis") # TRUE
not.exists("go") # FALSE
if(not.exists('hallo')) message("yes") # NULL
}
