% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.image.R
\name{not.image}
\alias{not.image}
\title{File name extension(s) is Not an image}
\usage{
not.image(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is not an image
}
\description{
Check if one or multiple file name entry is not an image
}
\examples{
img.1 = "fjk.jpg"
not.image(img.1)

img.2 = "fjk.bmp"
not.image(img.2)

img.3 = "fjk.SVG"
not.image(img.3)

# a vector of file names
v = c("logo.png", "business process.pdf",
"front_cover.jpg", "intro.docx",
"financial_future.doc", "2022 buybacks.xlsx")
(output = not.image(v))
}
