% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not.integer}
\alias{not.integer}
\title{Not an integer}
\usage{
not.integer(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is an integer
}
\description{
Check if entry is not an integer
}
\examples{
not.integer(23.43) # TRUE
not.integer(45L) # FALSE
if(not.integer(4L)) message("yes") # NULL

}
