% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.na.R
\name{not.na}
\alias{not.na}
\title{Not NA}
\usage{
not.na(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is NA
}
\description{
Check if entry is not NA
}
\examples{
not.na(NA) # FALSE
not.na(NULL) # logical(0)
if(not.na(45)) message("something") # TRUE

}
