% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.null.R
\name{not.null}
\alias{not.null}
\title{Not NULL}
\usage{
not.null(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is NULL
}
\description{
Opposite of is.null(). Check if entry is not NULL
}
\examples{
is.null("") # FALSE
not.null("") # TRUE
not.null(NULL) # FALSE
if(not.null(45)) message("something") # yes

}
