% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.numeric.R
\name{not.numeric}
\alias{not.numeric}
\title{Not numeric}
\usage{
not.numeric(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is numeric
}
\description{
Check if entry is not numeric
}
\examples{
not.numeric("45") # TRUE
not.numeric(45) # FALSE
if(not.numeric(45)) message("yes") # yes

}
