% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not.vector}
\alias{not.vector}
\title{Not a vector}
\usage{
not.vector(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is vector
}
\description{
Check if entry is not vector
}
\examples{
vect1 = list(r=1,t=3:10)
vect2 = LETTERS
not.vector(vect1) # FALSE
not.vector(vect2) # FALSE
if(not.vector(vect1)) message("yes") # NULL

}
