% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR.R
\name{or}
\alias{or}
\title{Alternative return for statements}
\usage{
or(test, alternative)
}
\arguments{
\item{test}{an object to return}

\item{alternative}{alternative object to return}
}
\description{
Return alternative if the value of expression is empty or NA or NULL
}
