% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{plus}
\alias{plus}
\title{Increment vector by value}
\usage{
plus(., add = 1)
}
\arguments{
\item{.}{vector of number(s)}

\item{add}{number to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Increment the content of a vector and resave as the vector
}
\examples{
num1 <- sample(330:400,10)
num1

# increment num1 by 1
num1 #before increment
inc(num1)
num1 #after increment

# increment num1 by 5
inc #before increment
inc(num1, add= 5)
inc #after increment
}
