% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{randString}
\alias{randString}
\title{Generate a random string}
\usage{
randString(n, length)
}
\arguments{
\item{n}{number of strings to create}

\item{length}{length of string to create}
}
\value{
one more random string of specific length
}
\description{
Create a random string of specified length
}
\examples{
# Task 1: create 1 random string string of length 5
randString(n = 1, length = 5)

# Task 2: create 5 random string string of length 10
randString(n = 5, length = 10)


# Task 3: create 4 random string string of length 16
randString(n = 4, length = 16)

}
