% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcolorconst.R
\name{rcolorconst}
\alias{rcolorconst}
\title{R Color Constant}
\usage{
rcolorconst(title = "R Color Constants")
}
\arguments{
\item{title}{title of the output}
}
\description{
R color constants
}
\examples{
rcolorconst()


}
