% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decommissionedPkg.R
\name{rdecomPkg}
\alias{rdecomPkg}
\title{Check whether an R package has been decommissioned in CRAN}
\usage{
rdecomPkg(package)
}
\arguments{
\item{package}{package name to query}
}
\description{
Check if an R package has been decommisioned in CRAN
}
\examples{
# check if package is decomissioned
rdecomPkg("cattonum")

}
