% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_combo_func1.R
\name{read.csv.print}
\alias{read.csv.print}
\title{Read in a CSV and print first X rows and Columns}
\usage{
read.csv.print(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  ...,
  dim = c(10L, 5L)
)
}
\arguments{
\item{dim}{dimension of CSV content to print}
}
\value{
read csv content
}
\description{
Read a dataset of type csv and print x rows and y columns
}
