% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_combo_func1.R
\name{read.table.print}
\alias{read.table.print}
\title{Read in a CSV and print first X rows and Columns}
\usage{
read.table.print(
  file,
  header = FALSE,
  sep = "",
  quote = "\\"'",
  dec = ".",
  numerals = c("allow.loss", "warn.loss", "no.loss"),
  row.names,
  col.names,
  as.is = !stringsAsFactors,
  na.strings = "NA",
  colClasses = NA,
  nrows = -1,
  skip = 0,
  check.names = TRUE,
  fill = NULL,
  strip.white = FALSE,
  blank.lines.skip = TRUE,
  comment.char = "#",
  allowEscapes = FALSE,
  flush = FALSE,
  stringsAsFactors = FALSE,
  fileEncoding = "",
  encoding = "unknown",
  text,
  skipNul = FALSE,
  dim = c(10L, 5L)
)
}
\arguments{
\item{dim}{dimension of table content to print}
}
\value{
read table content
}
\description{
Read a dataset of type csv and print x rows and y columns
}
