% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runl2}
\alias{runl2}
\title{Run R script in batch mode}
\usage{
runl2(file.in = NULL)
}
\arguments{
\item{file.in}{OPTIONAL. Input file, if absent, current opened file will be used.}
}
\description{
Run R script as a job and generate log file
}
