% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sd.gm}
\alias{sd.gm}
\title{Calculate geometric standard deviation and round}
\usage{
sd.gm(num, rm.na = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{rm.na}{remove NAs from the vector}

\item{round}{round result to decimal place}
}
\value{
the geometric standard deviation of a set of numbers
}
\description{
Calculate the geometric standard deviation
}
\examples{
num1 <- sample(330:400,10)
sd.gm(num1,rm.na=FALSE)

}
