% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.8.R
\name{sim.logistic}
\alias{sim.logistic}
\title{Generate logistic random values}
\usage{
sim.logistic(n, location = 0, scale = 1, min = 0, max = 1)
}
\arguments{
\item{n}{integer number of random values to generate}

\item{location}{numeric location parameter of the logistic distribution (default is 0)}

\item{scale}{numeric scale parameter of the logistic distribution (default is 1)}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
random values that follow a logistic distribution with the given parameters\cr
}
\description{
generates random values from a logistic distribution. It takes in the number of values to generate (n) and optional location and scale parameters.
}
\details{
Details:\cr\cr

- The function first generates uniform [0,1] random values using runif(n) \cr
- It then transforms these values using the quantile function qlogis() to map them to the logistic CDF\cr
- stats::qlogis() takes the uniform values and the location and scale parameters\cr
}
