% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.length}
\alias{sort.length}
\title{Sort vector by length of its content}
\usage{
\method{sort}{length}(vec, asc = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{asc}{TRUE or FALSE whether to sort by ascending or descending order}
}
\description{
Sort the length of the content of a vector
}
\note{
This function removes all NAs prior to sorting the vector
}
\examples{
# sort by length of content
x = c("acs","tt","jdssr","h","grab")
y = sort.length(vec = x)
z = sort.length(vec = x, asc = FALSE)


}
