% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sortLength}
\alias{sortLength}
\title{Sort vector by length of its content}
\usage{
sortLength(vec, asc = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{asc}{TRUE or FALSE whether to sort by ascending or descending order}
}
\value{
vector of items sorted by length
}
\description{
Sort the length of the content of a vector
}
\note{
This function removes all NAs prior to sorting the vector
}
\examples{
# sort by length of content
x = c("acs","tt","jdssr","h","grab")
sortLength(vec = x) # ascending order of length
sortLength(vec = x, asc = FALSE) # descending order of length


}
