% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_combo_func1.R
\name{strsplit.bool}
\alias{strsplit.bool}
\title{Split a string of values and return as boolean}
\usage{
strsplit.bool(
  x,
  split,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  type = 2
)
}
\arguments{
\item{type}{type of return, see the as.boolean function for more info}
}
\value{
boolean values based on split string
}
\description{
Given a sting, split by a separator into boolean
}
