% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_combo_func1.R
\name{strsplit.num}
\alias{strsplit.num}
\title{Split a string of numbers and return numeric}
\usage{
strsplit.num(x, split, fixed = FALSE, perl = FALSE, useBytes = FALSE)
}
\value{
numeric values based on split string
}
\description{
Given a sting, split by a separator into numbers
}
