% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envobj.R
\name{summarize.envobj}
\alias{summarize.envobj}
\title{Get all the environment objects and their sizes}
\usage{
summarize.envobj(envir = parent.frame())
}
\arguments{
\item{envir}{the environment to retrieve objects from}
}
\value{
a dataframe of all the variables within the environment
}
\description{
Retrieve the size contribution of all the available objects in the environment
}
\examples{
# Get a data frame of all environment objects and their size
summarize.envobj()
}
