% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquelen.R
\name{unique_len}
\alias{unique_len}
\title{Combine unique() and length()}
\usage{
unique_len(.)
}
\arguments{
\item{.}{object such as vector or names(dataframe)}
}
\value{
length of the unique items in a vector
}
\description{
Combine two frequently used function together to return the length of the unique items of a vector
}
\examples{
frenchnames1 = c("Léa","Obinna","Bastien","Léa","Obinna", "Hugo", "Amélie","Louis")
unique_len(frenchnames1)

}
