% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vector_pop}
\alias{vector_pop}
\title{Remove last n elements or specified elements from a vector like array_push in PHP}
\usage{
vector_pop(., n = 1, el = NULL)
}
\arguments{
\item{.}{first list}

\item{el}{vector to remove}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Shorthand to remove elements from a vector and save as the same name
}
\examples{
num1 <- sample(330:400,10)
num2 <-"rpkg.net"


num1 #num1 vector before pop
vector_pop(num1)
num1 #num1 vector after pop

}
