% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bool.R
\name{yesNoBool}
\alias{yesNoBool}
\title{Convert Yes/No to Binary or Logical}
\usage{
yesNoBool(
  ds,
  fldname,
  out = c("change", "append", "vector"),
  type = c("bin", "log")
)
}
\arguments{
\item{ds}{data frame}
}
\description{
Seemlessly convert a yes or no to either a binary or logical output
}
\details{
type - bin for binary, and log for logical
}
