% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add.header}
\alias{add.header}
\title{Snippet function to add header to a current opened file}
\usage{
add.header()
}
\value{
Inserts header content for file
}
\description{
Shorthand to add header
}
\examples{
if(interactive())
add.header()
}
