% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedR.R
\name{archivedPkg}
\alias{archivedPkg}
\title{Listing of all CRAN archived R packages}
\usage{
archivedPkg(
  startsWith = letters,
  inc.date = TRUE,
  as = c("data.frame", "vector")
)
}
\arguments{
\item{startsWith}{one letter that the package name starts with eg. a, e, f}

\item{inc.date}{should archive date be included in the result}

\item{as}{return result as data frame or as vector}
}
\value{
a data frame container listing of all archived R packages
}
\description{
Retrieve a list of all currently archived R packages
}
\note{
The startsWith argument should be one letter and should be in lowercase
}
\examples{
# Task: get archived R packages with names beginning with A
archivedPkg(startsWith = "a")

# Task: return the above packages as a vector
archivedPkg(startsWith = "a", as = "vector")



}
