% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedR.R
\name{archivedPkg}
\alias{archivedPkg}
\title{Listing of all CRAN archived R packages}
\usage{
archivedPkg(startsWith = letters)
}
\arguments{
\item{n}{number of packages to retrieve}
}
\value{
a data frame container listing of all archived R packages
}
\description{
Retrieve a list of all currently archived R packages
}
\note{
If the n argument is set to 0, all packages will be retrieved. Keep in mind that the larger the count or if all packages
are requested, this function becomes slower.
}
