% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedR.R
\name{archivedPkg}
\alias{archivedPkg}
\title{Listing of all CRAN archived R packages}
\usage{
archivedPkg(
  startsWith = c("all", letters),
  after = NULL,
  inc.date = TRUE,
  as = c("data.frame", "vector")
)
}
\arguments{
\item{startsWith}{one letter that the package name starts with eg. a, e, f}

\item{after}{packages archived after a specific date eg. 2011-05-10}

\item{inc.date}{should archive date be included in the result}

\item{as}{return result as data frame or as vector}
}
\value{
a data frame or vector containing listing of all archived R packages
}
\description{
Retrieve a list of all currently archived R packages and their archive date
}
\note{
The startsWith argument should be one letter and should be in lowercase
}
\examples{
# Task 1: get archived R packages with names beginning with A
head(archivedPkg(startsWith = "a"))


# Task 2: return the packages from Task 1 without including latest archive date
res.dt2 <- archivedPkg(startsWith = "b", inc.date = FALSE)
res.dt2[1:10,]

# Task 3: return the results from Task 2 as a vector
res.dt3 <- archivedPkg(startsWith = "c", inc.date = FALSE, as = "vector")
res.dt3[1:10,]

# Task 4: return the archived packages beginning with Y archived after 2022-08-12
# Note that startsWith should be lowercase

#without archive date
yRPkg <- archivedPkg(startsWith = "y", after= NULL)
nrow(yRPkg) #number of rows returned
head(yRPkg) #show first few rows

#with archive date
yRPkg2 <- archivedPkg(startsWith = "y", after= "2022-08-12")
nrow(yRPkg2) #number of rows returned
head(yRPkg2) #show first few rows

}
