% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compHist}
\alias{compHist}
\title{Compare histograms of two distributions}
\usage{
compHist(
  x1,
  x2,
  title,
  color = c("green", "black", "yellow"),
  xlab = "",
  ylab = "",
  separate = FALSE
)
}
\arguments{
\item{x1}{numeric distribution 1}

\item{x2}{numeric distribution 2}

\item{title}{title of the hostogram plot}

\item{color}{color vector for output}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{separate}{LOGICAL. whether to separate the plots}
}
\value{
return histogram comparison using basic histogram plot
}
\description{
For comparing values between two distributions.
Simply input the two distributions, and it generates a clear and
informative histogram that illustrates the differences between the data.
}
\details{
Users have the option to view individual histograms for each distribution
before initiating the comparison, allowing for a detailed examination of
each dataset's characteristics. This feature ensures a comprehensive
understanding of the data and enhances the user's ability to interpret
the results of the distribution comparison provided by this function.
}
\examples{
# compare two normal distributions

set.seed(123)
compHist(
  x1 = rnorm(1000, mean = 0),
  x2 = rnorm(1000, mean = 2),
  title = "Histogram of rnorm Distributions With Means 0 & 2",
  color = c("lightslateblue", "salmon", "mediumvioletred")
)

set.seed(123)
# separate the plots for preview
compHist(
  x1 = rnorm(1000, mean = 0),
  x2 = rnorm(1000, mean = 2),
  title = c("Plot Means 0", "Plot Means 2"),
  color = c("lightslateblue", "blue", "green"),
  separate = TRUE
)

}
