% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{const}
\alias{const}
\title{Mathematical constants}
\format{
An object of class \code{list} of length 10.
}
\usage{
const
}
\description{
List of mathematical functions
}
\note{
\cr 1. Pi (π) - Approximately 3.14159
\cr 2. Euler's Number (e) - Approximately 2.71828
\cr 3. Golden Ratio (φ) - Approximately 1.61803
\cr 4. Euler-Mascheroni Constant (γ) - Approximately 0.57721
\cr 5. Feigenbaum Constant (δ) - Approximately 4.66920
\cr 6. Champernowne Constant (C10) - Approximately 0.123456789101112...
\cr 7. Apéry's Constant (ζ(3)) - Approximately 1.20205
\cr 8. Gelfond-Schneider Constant (eπ) - Approximately 22.45915
\cr 9. Khinchin's Constant (K) - Approximately 2.68545
\cr 10. Ramanujan-Soldner Constant (μ) - Approximately 1.45136
}
\keyword{datasets}
