% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{const}
\alias{const}
\title{Mathematical constants}
\format{
An object of class \code{list} of length 2.
}
\usage{
const
}
\description{
List of mathematical functions
}
\details{
const list contains a list of mathematical constants below\cr\cr
- pi\cr
- e: Euler's number\cr
- golden_ratio\cr
- euler_mascheroni: Euler-Mascheroni constant\cr
- feigenbaum: Feigenbaum constant\cr
- champernowne: Champernowne constant\cr
- apery: Roger Apery's constant\cr
- gelfond_schneider: Gelfond-Schneider constant\cr
- khinchin: Aleksandr Khinchin constant\cr
- ramanujan_soldner: Ramanujan-Soldner constant\cr
}
\examples{
# definition of pi
const$defs$pi

# value of pi
const$values$pi


}
\keyword{datasets}
