% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{const}
\alias{const}
\title{Mathematical constants}
\format{
An object of class \code{list} of length 2.
}
\usage{
const
}
\description{
List of mathematical functions
}
\keyword{datasets}
