% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gm.R
\name{cv.gm}
\alias{cv.gm}
\title{Calculate geometric coefficient of variation and round}
\usage{
cv.gm(num, na.rm = TRUE, neg.rm = TRUE, pct = TRUE, round = 2)
}
\arguments{
\item{num}{vector of numbers}

\item{na.rm}{remove NAs from the vector}

\item{neg.rm}{remove negative values from the vector}

\item{pct}{TRUE or FALSE. should result be in percent}

\item{round}{round result to decimal place}
}
\value{
the geometric cv of a set of numbers
}
\description{
Calculate the coefficient of variation and round
}
\examples{
num1 <- sample(330:400,15)

#get geometric CV, represent as percent and round to 3 decimal places
cv.gm(num1,round = 3)

#or
geo.cv(num1,round = 3)

}
