% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersec.R
\name{getMyIP}
\alias{getMyIP}
\title{Cyber security functions}
\usage{
getMyIP()
}
\arguments{
\item{input}{input string containing IP}
}
\value{
a data frame containing the list of all connections

extract_IP returns a vector of IP addresses

local IP address numeric value
}
\description{
Get all active connections
}
\examples{
\donttest{
# Get a list of all connections on my system

gAc <- get_active_conn()
}

\donttest{
# Extract all IP addresses from a string

# Example 1
x = system("ipconfig /all", intern = TRUE)
extract_IP(x)

# Example 2
# This example demonstrates the separate
# extraction of an IP address one per vector element:
str1 = c("Two IP(66.544.33.54) addresses
were discovered in the scan.
One was at 92.234.1.0.",
"The other IP was 62.3.45.255.")
extract_IP(str1)

# Example 3:
# This example demonstrates the extraction
# of multiple IP addresses from a single vector element:
str2 = "Two IP addresses were discovered
in the scan. One was at 92.234.1.0.
The other IP was 62.3.45.255."
extract_IP(str2)
}

Get the IP address of the current user

Obtain my local IP address

\donttest{
# Fetch my IP address
getMyIP() # show my current IP address
}


}
