% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersec.R, R/getMyIP.R
\name{get_active_conn}
\alias{get_active_conn}
\alias{extract_IP}
\alias{getMyIP}
\title{Cyber security functions}
\usage{
get_active_conn()

extract_IP(input)

getMyIP()
}
\arguments{
\item{input}{input string containing IP}
}
\value{
a data frame containing the list of all connections

extract_IP returns a vector of IP addresses

local IP address numeric value
}
\description{
Get all active connections

Obtain my local IP address
}
\examples{

# Get a list of all connections on my system
get_active_conn()


# Extract all IP addresses from a string

# Example 1
x = system("ipconfig /all", intern = TRUE)
extract_IP(x)

# Example 2
# This example demonstrates the separate
# extraction of an IP address one per vector element:
str1 = c("Two IP(66.544.33.54) addresses
were discovered in the scan.
One was at 92.234.1.0.",
"The other IP was 62.3.45.255.")
extract_IP(str1)

# Example 3:
# This example demonstrates the extraction
# of multiple IP addresses from a single vector element:
str2 = "Two IP addresses were discovered
in the scan. One was at 92.234.1.0.
The other IP was 62.3.45.255."
extract_IP(str2)

# Fetch my IP address
getMyIP() # show my current IP address


}
