% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersec.R, R/getDate.R, R/getMyIP.R
\name{get_active_conn}
\alias{get_active_conn}
\alias{extract_IP}
\alias{getMyIP}
\title{Cyber security functions}
\usage{
get_active_conn()

extract_IP(input_string)

getMyIP()
}
\value{
a data frame containing the list of all connextions
}
\description{
Get all active connections

Obtain my IP address
}
\examples{

# Get all connections on my system
x = get_active_conn()


}
