% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyIP.R
\name{getMyIP}
\alias{getMyIP}
\title{Get the IP address of the current user}
\usage{
getMyIP()
}
\value{
local IP address numeric value
}
\description{
Obtain my local IP address
}
\examples{
\donttest{
# Fetch my IP address
getMyIP() # show my current IP address
}


}
