% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2week.R
\name{getWeekSeq}
\alias{getWeekSeq}
\alias{is.Date}
\alias{not.Date}
\alias{is.leap}
\title{Convert Dates into Numeric Week Counts}
\usage{
getWeekSeq(start_date, end_date, in.format = "\%m/\%d/\%y")

is.Date(x)

not.Date(x)

is.leap(yyyy)
}
\arguments{
\item{start_date}{A scaler of class Date (if this argument is populated, the date arg must be empty)}

\item{end_date}{A scaler of class Date; must be later than the start_date (if this argument is populated, the date arg must be empty)}

\item{in.format}{date input format}

\item{x}{date item to check}

\item{yyyy}{year numeric value eg 1989}
}
\value{
data frame of the dates along with their corresponding week
}
\description{
Convert the range of date to number of weeks
}
\examples{
getWeekSeq(start_date="12/29/25",end_date="1/8/26")

}
