% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_equations2.R
\name{pairDist}
\alias{pairDist}
\title{Calculate the distance of points from the center of a cluster}
\usage{
pairDist(data, n, round = NULL)
}
\arguments{
\item{data}{dataset with at least 2 columns}

\item{n}{Number of observations (scaler)}

\item{round}{round result to decimal place}
}
\value{
a named vector consisting of a row number and a pair-distance value
}
\description{
This function operates on multivariate data and
calculates the distance of points from the centroid of one or more clusters.
}
\section{Function utility}{

Used to generate the computations needed to model pair-distance measures in three dimensions
}

\examples{
n = 6
data = attenu[,1:2]

#basic example
pairDist(data)

# specify number of observations
pairDist(data, n)

# round results to 2 decimal points
pairDist(data, n,2)

}
\references{
the current function was adapted from one of the examples in the svgViewR package, \cr
https://cran.r-project.org/web/packages/svgViewR/svgViewR.pdf
}
