% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{is.lognormal}
\alias{is.lognormal}
\alias{is.normal}
\alias{is.uniform}
\alias{is.poisson}
\alias{is.gamma}
\alias{is.logistic}
\alias{is.weibull}
\title{Check if a data fits a Normal or LogNormal or Uniform or Poisson or Gamma or Logistic distribution}
\usage{
is.lognormal(values, alpha = 0.05, method = 1)

is.normal(values, alpha = 0.05, method = 1)

is.uniform(values, alpha = 0.05)

is.poisson(values, alpha = 0.05)

is.gamma(values, alpha = 0.05)

is.logistic(values, alpha = 0.05)

is.weibull(values, alpha = 0.05)
}
\arguments{
\item{values}{vector of values}

\item{alpha}{significance level to test p-value against}

\item{method}{method for calculation, where 1 = Shapiro-Wilk test and 2 = Kolmogorov-Smirnov test}
}
\value{
boolean value if lognormal distributed

boolean value if normal distributed

boolean value if uniform distributed

boolean value if poisson distributed

boolean value if gamma distributed

boolean value if logistic distributed

boolean value if logistic distributed
}
\description{
Check whether a vector of data contains values that fit a distribution
}
\details{
This function takes a numeric vector as its input. This vector contains the dataset that will be analyzed.
\cr\cr
\strong{For Normal and LogNormal:}\cr
This function first performs a Shapiro-Wilk test on the data to check if it comes from a normal distribution. The Shapiro-Wilk test checks if sample data came from a normal distribution. It returns a p-value, with a higher p-value indicating stronger evidence that the data is normally distributed.
If the p-value from the Shapiro-Wilk test is above a predefined threshold (such as 0.05), the data is considered normally distributed. In this case, the function returns "normal".
If the data fails the Shapiro-Wilk normality test, the function then checks if the data fits a lognormal distribution. It transforms the data by taking the natural logarithm of each value. It then runs the Shapiro-Wilk test on the transformed data.
If the p-value from this test on the logged data is above the threshold, the function returns "lognormal", indicating the untransformed data fits a lognormal distribution after logging.
If the data fails both the normal and logged Shapiro-Wilk tests, the function returns "neither", indicating the data does not appear to come from a normal or lognormal distribution based on the tests.
}
\note{
is.normal and is.lognormal uses the "Shapiro-Wilk test" from the utils package

is.uniform uses the "Kolmogorov-Smirnov test"

is.poisson uses the "Chi-squared test"

is.gamma uses the "Anderson-Darling test" or "Kolmogorov-Smirnov test"

is.logistic use the "Kolmogorov-Smirnov test"

is.weibull use the "Kolmogorov-Smirnov test"
}
\examples{

# Prepare all data to test
# Set the seed for reproducibility
set.seed(323)
lognormal_data <- stats::rlnorm(n = 1000, meanlog = 1, sdlog = 1) #lognormal data
normal_data <- stats::rnorm(n = 1000, mean = 10, sd = 3) #normal data
uniform_data <- stats::runif(10000,min=0,max=10) #uniform data
poisson_data <- stats::rpois(1000, lambda = 5) #poisson data
gamma_data <- stats::rgamma(1000,shape = 5, rate = 2) # gamma data
logis_data <- stats::rlogis(1000, location = 4, scale = 2)# Simulate logistic values
weibull_data <- stats::rweibull(1000, shape = 4, scale = 2) # weibull data

# EXAMPLE FOR is.lognormal

# Generate 1000 data points from a lognormal distribution with mean 0 and standard deviation 1

lognormal_data1 <- rlnorm(n = 100, meanlog = 0, sdlog = 1)
lognormal_data2 <- rlnorm(n = 10, meanlog = 0, sdlog = 1)
lognormal_data3 <- rlnorm(n = 5, meanlog = 0, sdlog = 1)


# Test if the data is lognormal
is.lognormal(lognormal_data)
is.lognormal(lognormal_data1)
is.lognormal(lognormal_data2)
is.lognormal(lognormal_data3)

# EXAMPLE FOR is.normal

# Set the seed for reproducibility
set.seed(1989)

# Generate 1000 data points from a normal distribution with mean 0 and standard deviation 1
normal_data1 <- rnorm(n = 500, mean = 0, sd = 1)
normal_data2 <- rnorm(n = 100, mean = 0, sd = 1)
normal_data3 <- rnorm(n = 10, mean = 0, sd = 1)

# Test if the data is normal
is.normal(normal_data)
is.normal(normal_data1)
is.normal(normal_data2)
is.normal(normal_data3)

# EXAMPLES for is.uniform

is.uniform(uniform_data)

\dontrun{
# EXAMPLE for is.poisson
is.poisson(poisson_data) # should be TRUE
is.poisson(normal_data) # should be FALSE
}
\dontrun{
# EXAMPLE for is.gamma

is.gamma(gamma_data) # check if it is gamma distribution
is.gamma(data.norm) # check if it is gamma distribution
}
\dontrun{
# EXAMPLE for is.logistic

is.logistic(logis_data) # test if it is logistic distribution
is.logistic(normal_data)
}

is.weibull(weibull_data) #should return TRUE
is.weibull(normal_data) #should return FALSE

}
