% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{is.lognormal}
\alias{is.lognormal}
\alias{is.normal}
\alias{is.uniform}
\alias{is.poisson}
\alias{is.gamma}
\alias{is.logistic}
\alias{is.weibull}
\alias{is.cauchy}
\title{Check if a data fits a Normal or LogNormal or Uniform or Poisson or Gamma or Logistic distribution}
\usage{
is.lognormal(values, alpha = 0.05, method = 1)

is.normal(values, alpha = 0.05, method = 1)

is.uniform(values, alpha = 0.05)

is.poisson(values, alpha = 0.05)

is.gamma(values, alpha = 0.05)

is.logistic(values, alpha = 0.05)

is.weibull(values, alpha = 0.05)

is.cauchy(values, alpha = 0.05)
}
\arguments{
\item{values}{vector of values}

\item{alpha}{significance level to test p-value against}

\item{method}{method for calculation, where 1 = Shapiro-Wilk test and 2 = Kolmogorov-Smirnov test}
}
\value{
boolean value if lognormal distributed

boolean value if normal distributed

boolean value if uniform distributed

boolean value if poisson distributed

boolean value if gamma distributed

boolean value if logistic distributed

boolean value if logistic distributed

boolean value if cauchy distributed
}
\description{
Check whether a vector of data contains values that fit a distribution
}
\details{
This function takes a numeric vector as its input. This vector contains the dataset that will be analyzed.
\cr\cr
\strong{For Normal and LogNormal:}\cr\cr
- Method 1: we perform the Shapiro-Wilk test on the (log-transformed) data to test for normality.
The null hypothesis of the Shapiro-Wilk test is that the data are normally distributed.
If the p-value is greater than the chosen significance level (typically 0.05),
we fail to reject the null hypothesis, indicating that the data may follow a log-normal distribution.\cr\cr
- Method 2: we perform the Kolmogorov-Smirnov test on the log-transformed data, comparing it to a normal distribution with the same mean and standard deviation. Again, if the p-value is greater than the chosen significance level, it suggests that the data may follow a log-normal distribution.
These tests provide a statistical assessment of whether your data follows a log-normal distribution.
}
\examples{

# Prepare all data to test
# Set the seed for reproducibility
set.seed(13200323)
lognormal_data <- stats::rlnorm(n = 1000, meanlog = 1, sdlog = 1) #lognormal data
normal_data <- stats::rnorm(n = 1000, mean = 10, sd = 3) #normal data
uniform_data <- stats::runif(10000,min=0,max=10) #uniform data
poisson_data <- stats::rpois(1000, lambda = 5) #poisson data
gamma_data <- stats::rgamma(1000,shape = 5, rate = 2) #gamma data
logis_data <- stats::rlogis(1000, location = 4, scale = 2)#logistic values
weibull_data <- stats::rweibull(1000, shape = 4, scale = 2) #weibull data
cauchy_data <- stats::rcauchy(1000, location = 4, scale = 2) #cauchy data

# EXAMPLE FOR is.lognormal

# Test if the data is lognormal
is.lognormal(lognormal_data)
is.lognormal(normal_data)
is.lognormal(uniform_data)
is.lognormal(poisson_data)
is.lognormal(gamma_data)
is.lognormal(logis_data)
is.lognormal(logis_data)

# EXAMPLE FOR is.normal

# Set the seed for reproducibility
set.seed(1989)

# Generate 1000 data points from a normal distribution with mean 0 and standard deviation 1
normal_data1 <- rnorm(n = 500, mean = 0, sd = 1)
normal_data2 <- rnorm(n = 100, mean = 0, sd = 1)
normal_data3 <- rnorm(n = 10, mean = 0, sd = 1)

# Test if the data is normal
is.normal(normal_data)
is.normal(normal_data1)
is.normal(normal_data2)
is.normal(normal_data3)

# EXAMPLES for is.uniform

is.uniform(uniform_data)

\dontrun{
# EXAMPLE for is.poisson
is.poisson(poisson_data) # should be TRUE
is.poisson(normal_data) # should be FALSE
}
\dontrun{
# EXAMPLE for is.gamma

is.gamma(gamma_data) # check if it is gamma distribution
is.gamma(data.norm) # check if it is gamma distribution
}
\dontrun{
# EXAMPLE for is.logistic

is.logistic(logis_data) # test if it is logistic distribution
is.logistic(normal_data)
}

is.weibull(weibull_data) #should return TRUE
is.weibull(normal_data) #should return FALSE

# EXAMPLES for is.is.cauchy

is.cauchy(uniform_data)

}
