% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.8.R
\name{is.lognormal}
\alias{is.lognormal}
\alias{is.normal}
\alias{is.uniform}
\alias{is.poisson}
\alias{is.gamma}
\alias{is.logistic}
\alias{is.weibull}
\title{Check if a data fits a Normal or LogNormal or Uniform or Poisson or Gamma or Logistic distribution}
\usage{
is.lognormal(values, alpha = 0.5)

is.normal(values, alpha = 0.5)

is.uniform(values, alpha = 0.5)

is.poisson(values, alpha = 0.5)

is.gamma(values, alpha = 0.5)

is.logistic(values, alpha = 0.5)

is.weibull(values, alpha = 0.5)
}
\arguments{
\item{values}{vector of values}

\item{alpha}{significance level to test p-value against}
}
\value{
boolean value if lognormal distributed

boolean value if normal distributed

boolean value if uniform distributed

boolean value if poisson distributed

boolean value if gamma distributed

boolean value if logistic distributed

boolean value if logistic distributed
}
\description{
Check whether a vector of data contains values that fit a distribution
}
\details{
This function takes a numeric vector as its input. This vector contains the dataset that will be analyzed.
\cr\cr
\strong{For Normal and LogNormal:}\cr
This function first performs a Shapiro-Wilk test on the data to check if it comes from a normal distribution. The Shapiro-Wilk test checks if sample data came from a normal distribution. It returns a p-value, with a higher p-value indicating stronger evidence that the data is normally distributed.
If the p-value from the Shapiro-Wilk test is above a predefined threshold (such as 0.05), the data is considered normally distributed. In this case, the function returns "normal".
If the data fails the Shapiro-Wilk normality test, the function then checks if the data fits a lognormal distribution. It transforms the data by taking the natural logarithm of each value. It then runs the Shapiro-Wilk test on the transformed data.
If the p-value from this test on the logged data is above the threshold, the function returns "lognormal", indicating the untransformed data fits a lognormal distribution after logging.
If the data fails both the normal and logged Shapiro-Wilk tests, the function returns "neither", indicating the data does not appear to come from a normal or lognormal distribution based on the tests.
}
\note{
is.normal and is.lognormal uses the "Shapiro-Wilk test" from the utils package

is.uniform uses the "Kolmogorov-Smirnov test"

is.poisson uses the "Chi-squared test"

is.gamma uses the "Anderson-Darling test" or "Kolmogorov-Smirnov test"

is.logistic use the "Kolmogorov-Smirnov test"

is.weibull use the "Kolmogorov-Smirnov test"
}
\examples{
# EXAMPLE FOR is.lognormal

Set the seed for reproducibility
set.seed(1989)

# Generate 1000 data points from a lognormal distribution with mean 0 and standard deviation 1
lognormal_data <- rlnorm(n = 1000, meanlog = 0, sdlog = 1)
lognormal_data1 <- rlnorm(n = 100, meanlog = 0, sdlog = 1)
lognormal_data2 <- rlnorm(n = 10, meanlog = 0, sdlog = 1)
lognormal_data3 <- rlnorm(n = 5, meanlog = 0, sdlog = 1)

# Plot the distribution of the data using a histogram
hist(lognormal_data, main = "Lognormal Distribution with Mean 0 and Standard Deviation 1", xlab = "Data Values", col = "lightblue")

# Add a density curve to the histogram
lines(density(lognormal_data), col = "red")

# Test if the data is lognormal
is.lognormal(lognormal_data)
is.lognormal(lognormal_data1)
is.lognormal(lognormal_data2)
is.lognormal(lognormal_data3)

# EXAMPLE FOR is.normal

# Set the seed for reproducibility
set.seed(1989)

# Generate 1000 data points from a normal distribution with mean 0 and standard deviation 1
normal_data <- rnorm(n = 1000, mean = 0, sd = 1)
normal_data1 <- rnorm(n = 500, mean = 0, sd = 1)
normal_data2 <- rnorm(n = 100, mean = 0, sd = 1)
normal_data3 <- rnorm(n = 10, mean = 0, sd = 1)

# Plot the distribution of the data using a histogram
hist(normal_data, main = "Normal Distribution with Mean 0 and Standard Deviation 1", xlab = "Data Values", col = "lightblue")

# Add a density curve to the histogram
lines(density(normal_data), col = "red")

# Test if the data is normal
is.normal(normal_data)
is.normal(normal_data1)
is.normal(normal_data2)
is.normal(normal_data3)

# EXAMPLES for is.uniform

set.seed(323)
unifdata <- runif(10000,min=0,max=3)

is.uniform(unifdata)


# EXAMPLE for is.poisson
set.seed(1989)
data.poisson <- rpois(1000, lambda = 5) #poisson data
data.normal <- runif(1000, min = 5, max = 10) #uniform data
is.poisson(data.poisson) # should be TRUE
is.poisson(data.normal) # should be FALSE


# EXAMPLE for is.gamma

set.seed(5434)
data.gamma <- stats::rgamma(1000,shape = 5, rate = 2) # gamma data
data.norm = runif(1000,min = 0, max = 1) # not gamma data

is.gamma(data.gamma) # check if it is gamma distribution
is.gamma(data.norm) # check if it is gamma distribution


# EXAMPLE for is.logistic

set.seed(231)
n <- 1000
location <- 4
scale <- 2
xlogi <- sim.logistic(n, location, scale)# Simulate logistic values
hist(xlogi, prob=TRUE, main="Plot of simulated logistic")# Plot histogram

data.normal <- runif(1000) #uniform ditribution for negative test

is.logistic(xlogi) # test if it is logistic distribution
is.logistic(data.normal)

data.weibull <- rweibull(1000, shape = 4, scale = 2) # weibull data
data.norm = runif(1000, min = 0, max = 1) # not weibull data

is.weibull(data.weibull) #should return TRUE
is.weibull(data.norm) #should return FALSE

}
