% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{#' @export
#' Next version}
\usage{
x \%!in\% table
}
\description{
sample_by_column <- function(.dt, col, n) {
  .dt[.dt[, as.character(substitute(col))] %in% sample(unique(.dt[, as.character(substitute(col))]), n), ]
}
}
\details{
#' @export
#' Next version

has <- function(., var, col, row) {
  stop("Not completed")

  object <- .

  if (is.data.frame(object)) {
    # check if var exists in data.frame

    # check if column exists in data.frame

    # check if row exists in data.frame
  }

  if (is.vector(object)) {
    # check if var exists in vector
  }

  if (is.character(object)) {
    # check if var exists in character
  }

  if (is.numeric(object)) {
    # check if var exists in number
  }
}
}
