% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.image.R
\name{is.image}
\alias{is.image}
\title{Is item an image}
\usage{
is.image(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is an image
}
\description{
Check if entry is an image
}
\examples{
img.1 = "fjk.jpg"
is.image(img.1)

img.0 = "fjk.bbVG"
is.image(img.0)

img.2 = "fjk.bmp"
is.image(img.2)

img.3 = "fjk.SVG"
is.image(img.3)

}
