% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_push}
\alias{list_push}
\title{Add elements to a list like array_push in php}
\usage{
list_push(., add)
}
\arguments{
\item{.}{first list}

\item{add}{list to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Shorthand to add elements to a vector and save as the same name
}
\examples{
num1 <- list(sample(330:400,10))
num2 <-list("rpkg.net")
list_push(num1, add= num2)
}
