% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_filter.R
\name{mutate_filter}
\alias{mutate_filter}
\title{Mutate a subset of dataset, and reattach it to the dataset}
\usage{
mutate_filter(., subset, ...)
}
\arguments{
\item{subset}{subset of data to modify}

\item{...}{mutation syntax similar to dplyr::mutate}
}
\value{
data frame containing original data, but with a subset mutated
}
\description{
Extension of the dplyr::mutate function that allows the user to mutate only a specific filtered subset of a data, while leaving the other parts of the data intact
}
\examples{
#mutate a subsection filter of mtcars
dt =mtcars
mutate_filter(dt,mpg == 21.0 & cyl == 6, cyl=1000,hp=2000,vs=hp*2)
}
