% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_filter.R
\name{mutate_filter}
\alias{mutate_filter}
\title{Mutate a subset of dataset, and reattach it to the dataset}
\usage{
mutate_filter(., filter., ...)
}
\arguments{
\item{filter.}{subset of data to modify}

\item{...}{mutation syntax similar to dplyr::mutate}
}
\description{
Mutate a subset of dataset, and reattach it to the dataset
}
