% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in vector or array}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector entry}

\item{table}{table of items to check}
}
\value{
a boolean value to indicate if entry is present
}
\description{
Check if entry is in vector
}
\examples{
5 \%nin\% c(1:10) #FALSE
5 \%nin\% c(11:20) #TRUE

x = "a"
if(x \%nin\% letters) x
}
