% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.image.R
\name{not.image}
\alias{not.image}
\title{Not an image}
\usage{
not.image(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is not an image
}
\description{
Check if entry is not an image
}
\examples{
img.1 = "fjk.jpg"
not.image(img.1)

img.2 = "fjk.bmp"
not.image(img.2)

img.3 = "fjk.SVG"
not.image(img.3)

}
