% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not.null}
\alias{not.null}
\title{Not NULL}
\usage{
not.null(x)
}
\arguments{
\item{x}{vector entry}
}
\value{
a boolean value to indicate if entry is NULL
}
\description{
Check if entry is not NULL
}
\examples{
not.null("") # TRUE
not.null(NULL) # FALSE
if(not.null(45)) message("yes") # yes

}
