% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{number}
\alias{number}
\title{Generate a random number}
\usage{
number(n)
}
\arguments{
\item{n}{how many numbers to generate}
}
\value{
random numbers between 1 and 1 billion
}
\description{
Shorthand code to generate a random number
}
\examples{
number(1)
number(10)
}
