% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcolorconst.R
\name{rcolorconst}
\alias{rcolorconst}
\title{R Color Constant}
\usage{
rcolorconst(title = "R Color Constants")
}
\arguments{
\item{title}{title of the output}
}
\value{
returns color constant
}
\description{
R color constants description
}
\details{
Additional details...
}
\section{Use case}{

}

\examples{
# Without title
rcolorconst()

# With title
rcolorconst("My new color constant")

}
