% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.7_combo_func1.R
\name{strsplit.num}
\alias{strsplit.num}
\title{Split a string of numbers and return numeric}
\usage{
strsplit.num(x, split, fixed = FALSE, perl = FALSE, useBytes = FALSE)
}
\arguments{
\item{x}{character vector, each element of which is to be split. Other inputs, including a factor, will give an error.}

\item{split}{character vector}

\item{fixed}{logical. If TRUE match split exactly, otherwise use regular expressions. Has priority over perl.}

\item{perl}{logical. Should Perl-compatible regexps be used?}

\item{useBytes}{logical. If TRUE the matching is done byte-by-byte rather than character-by-character, and inputs with marked encodings are not converted.}
}
\value{
numeric values based on split string
}
\description{
Given a sting, split by a separator into numbers
}
\examples{
# to be added
}
