% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_push.R
\name{vector_push}
\alias{vector_push}
\title{Add elements to a vector like array_push in PHP}
\usage{
vector_push(., add)
}
\arguments{
\item{.}{first vector}

\item{add}{vector to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Shorthand to add elements to a vector and save as the same name
}
\examples{
num1 <- sample(330:400,10)
num1 #preview
num2 <-"rpkg.net"


#seamlessly add num2 to num1 and re-save as num1

vector_push(num1,num2)
num1 #updated with num2
num2 #not updated

}
