% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vector_push}
\alias{vector_push}
\title{Calculate elements to a vector like array_push in php}
\usage{
vector_push(., add)
}
\arguments{
\item{.}{first vecotr}

\item{add}{vector to add}
}
\value{
vector combining fist and second vector, but have name set to the first
}
\description{
Shorthand to add elements to a vector and save as the same name
}
\examples{
num1 <- sample(330:400,10)
num2 <-"rpkg.net"
vector_push(num1, add= num2)
}
