% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vector_shuffle}
\alias{vector_shuffle}
\title{Shuffle a vector just like shuffle in php}
\usage{
vector_shuffle(., replace = FALSE, prob = NULL)
}
\description{
Shuffle a vector just like shuffle in php
}
