% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_shuffle.R
\name{vector_shuffle}
\alias{vector_shuffle}
\title{Shuffle a vector just like shuffle in PHP}
\usage{
vector_shuffle(., replace = FALSE, prob = NULL, seed = NULL)
}
\arguments{
\item{.}{vector to shuffle}

\item{replace}{replace selected value}

\item{prob}{probability of occurrence}

\item{seed}{apply seed if indicated for reproducibility}
}
\value{
shuffled vector of items store to the vector name
}
\description{
Shorthand to shuffle a vector and save
}
\examples{
v1<-c(3,45,23,3,2,4,1)


#demonstrate vector_shuffle
vector_shuffle(v1)
v1 # show outputs

#demonstrate reproducibility in shuffle with seed
v0<-v1
vector_shuffle(v0)
v0 #first output

v0<-v1
vector_shuffle(v0)
v0 # different output from first output top

v0<-v1
vector_shuffle(v0,seed = 232L)
v0 #second output

v0<-v1
vector_shuffle(v0,seed = 232L)
v0 #the same output as second output top
}
