% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bool.R
\name{yesNoBool}
\alias{yesNoBool}
\title{Convert Yes/No to Binary or Logical}
\usage{
yesNoBool(
  table,
  fldname,
  out = c("change", "append", "vector"),
  type = c("bin", "log")
)
}
\arguments{
\item{table}{data frame}

\item{fldname}{field name in the data frame}

\item{out}{output form, choices - change, append, vector}

\item{type}{output type, choices - bin, log}
}
\description{
Seemlessly convert a yes or no to either a binary or logical output
}
\details{
type - bin for binary, and log for logical
}
\examples{
usedata <- mtcars
usedata
usedata$yess = "yes"
yesNoBool(usedata,"yess")


}
