% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsore.R
\name{zscore}
\alias{zscore}
\title{Calculates Z-Scores of a distribution}
\usage{
zscore(data, round)
}
\arguments{
\item{data}{dataiable}

\item{round}{round output to how many decimal place description}
}
\description{
Calculates Z-Scores based on data
}
\examples{
Capture z-scores from the following distribution x
x = c(6, 7, 7, 12, 13, 13, 15, 16, 19, 22)
z_scores = zscore(x, round = 2) # limit to 2 decimal place
z_scores = zscore(x) # no decimal place limit

df = data.frame(val = x, zscore = z_scores)
head(df)
}
