% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsore.R
\name{zscore}
\alias{zscore}
\alias{ZscoreGrowthCurve}
\title{Calculates Z-Scores of a distribution}
\usage{
zscore(data, round, na.rm = TRUE)

ZscoreGrowthCurve(X, M, S, L = !0)
}
\arguments{
\item{data}{dataiable}

\item{round}{round output to how many decimal place description}

\item{na.rm}{remove NA values before calculating z-scores}

\item{X}{description}

\item{M}{description}

\item{S}{description}

\item{L}{description}
}
\description{
Calculates Z-Scores based on data
}
\examples{
Capture z-scores from the following distribution x
x = c(6, 7, 7, 12, 13, 13, 15, 16, 19, 22)
z_scores = zscore(x, round = 2) # limit to 2 decimal place
z_scores = zscore(x) # no decimal place limit

df = data.frame(val = x, zscore = z_scores)
head(df)
# Calculate the zscore for a patient weighing 50kg
L=-0.1600954
M=9.476500305
S=0.11218624
X=50
ZscoreGrowthCurve(X,M,S,L)

}
\references{
CDC growth chart Z score calculation: https://www.cdc.gov/growthcharts/zscore.htm
}
